+++ tests/spec/modular/http-client.ts
@@ -21,9 +21,9 @@ interface ApiConfig<SecurityDataType> {
   securityWorker?: (securityData: SecurityDataType) => RequestParams;
 }
 
-interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
-  data: D | null;
-  error: E | null;
+export interface HttpResponse<D extends unknown, E extends unknown = unknown> extends Response {
+  data: D;
+  error: E;
 }
 
 export enum BodyType {
@@ -97,8 +97,8 @@ export class HttpClient<SecurityDataType = unknown> {
 
   private safeParseResponse = <T = any, E = any>(response: Response): Promise<HttpResponse<T, E>> => {
     const r = response as HttpResponse<T, E>;
-    r.data = null;
-    r.error = null;
+    r.data = (null as unknown) as T;
+    r.error = (null as unknown) as E;
 
     return response
       .json()